Revert logic for determining if circle present
The purpose of !angular_travel is specifically to ensure that zero cases
with different target and initial positions get handled at the top
rather than at the bottom where they get handled inconsistently. The
added parantheses changed the logic back to what it was initially, with
zeros with different targets getting handled by the switch statement.