Z_SENSORLESS has confused error message
The default error message "SENSORLESS_PROBING requires a TMC stepper driver with StallGuard on Z."  is confusing. 

The error does not indicate the source of the problem (it implies something non-specific is amiss with the StallGuard configuration) ..
However it turns out that Z_SENSORLESS (which was defined) is being implicitly disabled by the USE_PROBE_FOR_Z_HOMING which is implicitly enabled by Z_MIN_PROBE_USES_Z_MIN_ENDSTOP_PIN 

Ashamed to admit how long it took me to figure out why the error was occurring.