MightyBoard: Remove MegaCore and add optimized build flags
The MegaCore library not only defines pins differently than the extended library (and the old Sailfish instructions for modifying the pins file), but it also introduces a number of different library versions that will complicate debugging efforts. Now that the mega2560ext variant is here and tested to be stable, this will switch the MightyBoard over to using the variant definition provided by that without also requiring different library versions.

Additionally, this commit / PR adapts the optimizations / tuning previously introduced for the atmega1284p into use for both the atmega2560 and atmega1280 on the Mightyboard. These build flags are stable on these chips and will reduce the code size considerably.