Completely move away from DWIN driven navigation
Because DWIN is a complete PoS, it isn't always consistent in what it reports on the PIC_NOW command. So, sometimes it says to the firmware "Oh, I'm on the print running screen" but it is actually displaying the "Print tuning screen". So the firmware then receives all kinds of events it doesn't recognize because the display is telling it is on a different screen!