Prevent hot-end idle timeout settings from interfering with PROBING_HEATERS_OFF
Previously the HOTEND_IDLE_TIMEOUT was (ab)used for disabling the heaters, but HOTEND_IDLE_TIMEOUT doesn't necessarily turn off the heaters. It just may set them to a lower temperature. In addition, some people like to level at a high temperature and we do not want to reset their temperature.

Fixes #87 also