I like it better the way it was before
I put back the code that was previously removed because I would rather have the option to use an external library or not to use an external library.   This version of the PR allows build_flags to be used which will load a user library for MAX31855 or a user library for MAX31865 or a user library for MAX6675 or use the Adafruit MAX31865 V1.1.0 library or NOT to use a external library for MAX31855 or NOT to use an external library for MAX6675.  If an external library is not used then the indivdual frameworks will minipulate the PINS like they did before (like my PR never happend).  The following buid_flags have been implemented:  -DLIB_MAX31865=1 means to load the external library in platformio.ini defined by LIB_USR_MAX31865 = '...' If -DLIB_MAX31865=1 is missing from the build_flags the Adafruit MAX31865 V1.1.0 library will be loaded (LIB_ADAFRUIT_MAX31865=Adafruit MAX31865 library@~1.1.0). If -DLIB_MAX31855=1 is included in the build_flags then an external user library defined in platformio.ini file with LIB_USR_MAX31855 = '...' is used for the MAX31855. If -DLIB_MAX31855 is missing from the build_flags then the pervious method of collecting the data will be used (the HAL will manipulate the SS_PIN and call max6675_spi.init() if it needs to and receive data via max6675_spi.receive(), or spiRec()).  If the -DLIB_MAX6675=1 is included in the build_flags then the user library defined in platformio.ini file by LIB_USR_MAX6675='...' will be used for MAX6675. If the -DLIB_MAX6675=1 is missing from the build_flags then the previous way of collecting data for the MAX6675 sensor will be used.  In addition to these build_flag for user libraries additional flags were added for handling the pin mapping bug. These additional flags are only used by the user defined library: -DLARGE_PINMAP=1 will allow the object initialization to occur with 'uint32_t' for the PIN VALUES; -DTEMP_MODE=3 will allow the hardware SPI mode to change to SPI_MODE3; -DTEMP_MODE=2 will allow the hardware SPI mode to change to SPI_MODE2; -DTEMP_MODE=1 will allow the hardware SPI mode to change to SPI_MODE1 and -DTEMP_MODE=0 will allow the hardware SPI mode to change to SPI_MODE0.  To accomodate the previous method of collecting data for MAX31855 and MAX6675 a compatibility layer was added incase the CS_PIN was used.  The previous method only used the SS_PIN for enabling and disabling the collection of data.  Even tho this is more complex a method, at least it will be up to the user to either create their own library or use the default method that Marlin uses.  At least this PR will stop the manipulation of SS_PIN only when it is really needed