Prevent rapid adjustment of settings causing a factory reset - fixes #123
Apparently doing rapid Z-offset adjustments and subsequently saving settings is _SLOW_. So slow, that Marlin is not able to tell the watchdog it is still alive. So the watchdog loses patience and resets the CPU. This also corrupts the EEPROM in the process.

To fix it, now only save the EEPROM in a 250ms loop when requested.