Consider axes unhomed after M84
When the steppers are released/disabled on a M84, after for example a time-out or manual disable command. They should not be trusted anymore.

Example, when you do an auto bedleveling, wait after the timeout, move your bed and restart another auto bedleveling, you'll get in trouble without a new autohome.
After a M84, the Z axis also drops a bit and that alone should be untrusted.

The best place I could think of was in MarlinCore.cpp, I have tested this on a E3v2 and the result was as desired.