Display Extra Info during Probing Screen & change probe speed
I was able to change the probing screen so that it will now show the last Z measurement, as well as the the # of good points found. This way if it is constantly resetting, the user has some feedback. (It had entered an infinite loop for me once due to an uneven bed, and there was no feedback when it reset the count)