Start implementing a generic serial hook class. This is currently not finished.
The idea is to have the compiler optimize all the hook functions when none are being used without relying on a preprocessor configuration hell.
All of this code will be optimized away when built so it should not impact the binary size nor the runtime speed.
Once finished, it should just work as expected while simplifying the current serial.h file.

This will allow to have as many serial port as desired, and, hopefully, should also permit to remove many redundant code in the different HAL, while hooking any serial port (for GUI console or telnet or whatever).