Add option `WAIT_FOR_HOTEND` under `PROBING_HEATERS_OFF`.
Wait for the hotend to come back up to temp between probes, to ensure that the temperature
does not droop excessively when probing. This was pointed out in #20595 - when running a
long `G29` command before printing, but after a `M109`, the planner will trudge ahead at
whatever temperature the nozzle happens to be at after probing. In other cases this can
be useful if one wants to have a constant nozzle temp when probing, due to thermal expansion,
such as when using `NOZZLE_AS_PROBE` or anything else that uses the nozzle directly.