Optimize least_squares_fit
Remove the z2bar variable (which is unused) and don't bother storing
intermediate values back into the linear_fit_data struct in function
finish_incremental_LSF().

Saves 232 bytes of program memory on AVR.

This function could be optimized for speed by calculating (1/N) once
and using multiplications instead of divisions, but this code is not
speed-critical and that would increase program memory by a few bytes.