Add cheap implementations of strtol() and strtof()
We don't need all functionality of strtol() (we always use base-10 and
never check for errors) and for strtof() we even explicitly disable
the exponent functionality. So if we implement our own functions
instead of importing from the standard library we can gain a lot of
space/speed.

This saves 1516 bytes of program memory on AVR and speeds up parsing of
floats by around 50% and parsing of ints by around 75%.