Generalize MMU with variable nr of colors
There are more Prusa MMU compatible devices on the wild and a more general approch as been taken for such devices.
SMuff V5 has been also discontinued then less user may use it in the future but on the other side Enraged Rabbit Carrot Feeded is growing up and "should" be used with SMuFF firmware.