[RFC][WIP] Remove usage of sprintf()
Using sprintf() might make the code simpler to read but it has a big
cost in program memory and runtime speed.

I propose this approach with simpler functions that keep track of the
end of the buffer we're printing to by using [ui][l]toa() and dtostrf()
functions.

The non-printf code is faster and saves about 1.5Kb of program memory
in my Anet A6.