bug: Invalid PlatformIO option for memory offset
A subset of STM32F1 & STM32F4 board definitions are using the option
`board_build.address` to specifying the memory location for firmware.

The correct option to be used is `board_upload.offset_address`, which
was added in [platformio/platform-ststm32#853208d1][1].  PR #21488
(recently submitted) uses the correct option for some _newly_ added
hardware abstractions, but left this in place for existing hardware.

Using this option presently overwrites the bootloader when using
`openocd` and an STLinkV2.

For example: https://ask.csdn.net/questions/4185515

This is occurring because the option gets ignored and the default vector
table location of `0x08000000` gets used (which is the location of the
bootloader).

NOTE: All changes were made using the following mechanism

```bash
sed -i 's/board_build.address/board_upload\.offset_address/g' ini/stm32f{1,4}.ini
```

No other changes to spacing or validation beyond verification on the
STM32F103RE_btt board.  This is confirmed operational with the offset
previously used with the invalid option, `0x08007000`.

[1]: https://github.com/platformio/platform-ststm32/commit/853208d1