Allow individual baudrates for each TMC driver
This patch allows for each axis with a TMC driver a different baudrate to be
used. This is done by defining the macro TMC_<axis>_BAUDRATE to the desired
baudrate, for example: #define TMC_X_BAUDRATE 250000 to define the baudrate for
the X-axis TMC stepper driver.
If a baudrate for a stepper is not defined in this per-axis fashion, the
baudrate defined in the general TMC_BAUD_RATE macro is used instead. Thus,
default behaviour is kept.