Fix POWER_LOSS_PIN redefinition warning
If POWER_LOSS_PIN is defined in Configuration_adv.h the compiler rises a warning about POWER_LOSS_PIN redefinition. This means that currently the POWER_LOSS_PIN configured is not able to override the default one.

The worst case is if the custom POWER_LOSS_PIN is set to -1. In that case the printer believes to be always in power loss condition,

This fixes that issue.