revert to (more or less) the first implementation of the PTC/BTC float checks
The checks as-is are failing in a number of ways. First, some typos.. _test_*
variables no longer exist, PTC_SAMPLE_END should be PTC_SAMPLE_RES. After
fixing those, the checks are not actually working (floats are passed OK). This
brings back the 'auto' vars that would be float if the user used '30.0f' or
the like; the type from the auto var is copied to a known-float & ensured that
the known value doesn't stay a float (i.e. gets rounded to an int: 12).