Triple Endstop bugfix
Seems the triple endstop section was missing some information. I can't be certain I did this correctly, I don't fully understand what the added info does, but it compiles and works. I just used what was in the dual endstop section to fill in the missing info.

I verified this on a SKR Pro, with actual hardware.