Move zpounce after restored current to avoid skipped steps coming off of the gantry.
I use G34 on a corexy with 3 Z-steppers on one driver. 
the zpounce here skips after when using a very low current for the zgrind. 
in my set up, zgrind is lowering the bed so the current requirement is very low. 
zpounce is physically raising the bed and always skips steps here.