MightyBoard_RevE fix EX2-HEAT, EX2-FAN MOSFET pins
As per issue #23956, the two MOSFET pins for the EX2-HEAT and EX2-FAN were erroneously swapped thereby causing a safety issue by enabling heat on EX2 when instead the fan should have turned on. This can happen because both extruders in a Replicator 2, FFCP, or clone share a heat break and therefore there will be some heating of EX2 when EX1 is in use.