Mightyboard - Fix erroneous LED_PIN definition
Per the Mightyboard Rev.E schematic, the assignment of LED_PIN to MOSFET_6_PIN is incorrect. The RGB controller does not have a direct power cutoff, and the closest simple match (using the pin number still in the code comment) is the 'BLINK' line leading to onboard LED DEBUG1. This PR properly reassigns logical pin 13 to the LED_PIN define, restoring the correct and intended operation.