PREVENT_COLD_EXTRUSION from LCD
Make users enable/disable cold extrusion directly from LCD with #define COLD_EXTRUSION_MENU_ITEM in Configuration.h

This is a feature that I've actually had a lot of request for, when compiling firmware for friends and colleagues from work.

This feature makes it easy to calibrate e-steps, without connecting printer to a terminal. Specially for bowden-systems, where filament only have to run through extruder, and not near hotend.

This feature is tested and works perfectly with LCD12864.