Update pins_TENLOG_D3_HERO.h
#define PS_ON_PIN          40

Pin 40 is the power supply unit pin for M81 gcode command if your tenlog v2.1, v2.2 or v2.3 mainboard supports it. 
Some tenlog boards will auto power off on startup if the pin is not defined. When you flip the power switch, the printer powers up for a moment and then powers off.
I reccomend this change should be commented out by default as I'm not sure it is applicable for all tenlog boards.