On pause still report "SD printing byte X/Y".
When printing from sd card and print is paused (via lcd or filament
runout happens) then Marlin starts reporting "Not SD printing" which
is not really what is happening.

This causes problem with OctoPrint. When it sees "Not SD printing"
is considers that printing is cancelled and migrates it's state
from "Pausing" to "Cancelling", executes cancel-gcode (defined by user)
etc and basically stops entire print. (In my cause that was on filament
runout).

Anyway, reporting "SD printing byte X/Y" (where X simply doesn't
raise) makes sense. We are indeed in middle of printing.