Add M524 to emergency parser to allow fast sd print stopping.
Currently there is no equivalent of "Stop print" (from LCD) that can be
run remotely and get *fast* printing stop. That's a problem
if you observe that something is not going to work with print and want
to stop as fast as possible (but not to kill everything like M112).

Add M524 (abort sd print) to emergency parser which does the job.