- reliability upgrade for ATmega2560 MCU series (640-1280-1281-2560-2561)
The _ATmega_savePinAlternate function saves the device state which affects the provided GPIO pin and disables it. This way the provided pin can be safely used for GPIO purposes until the _ATmega_restorePinAlternate function is called. Then the device state is reset prior to the call of _ATmega_savePinAlternate, assuming that no other device state change related to the disabled peripherals was undertaken inbetween the function calls.

- added REPRAP_FULL_GRAPHIC_SMART_CONTROLLER to MKS TinyBee V1.0