add code to allow PA13 to be used as general purpose I/O
The Trigorilla board uses PA13 to drive the case led.

Currently it can't be used as general purpose I/O because the default state has it dedicated to the JTAG/SWD port.

This code sets bit 26 in the AFIO_MAP register to disable the JTAG/SWD port.  It then sets PA13 as an output.