Move bltouch._reset() after the move to a safe height
If it is triggered, something is triggering it.  This is about to move
up to a safe z height anyway, which could clear the trigger condition.
The move will also give it additional time to settle if it is
transient.

Remove only checking if it is in high_speed_mode, while it should only
be deployed if in high speed mode, if it is in an alarm condition it
needs to be reset either way.