eeprom save fix
On TMC2208_STANDALONE with AUTO_BED_LEVELING_BILINEAR and GRID_MAX_POINTS_X greater 7, so as of 8x8 a crc mismatch occurs on reboot. Mesh and other settings can be generated and saved but on printer reboot everything is resetted.

Error:EEPROM CRC mismatch - (stored) 57991 != 44219 (calculated)!

UBL works still. It only affects AUTO_BED_LEVELING_BILINEAR as of 8x8.

With this fix the problem is solved.