G2/G3: Fix Z/helix planning on UBL
When interpolating the Z (and other non-XY) axes, reconstruct the
position from start of the arc and current segment index, *then* apply
the modifiers, so that modifiers are not accumulated for each vertex.

Recalculating each absolute vertex location also reduces the error (and
resulting jitter) of the Z/E axis for long moves at high resolution. For
the same reason, this is also now done uniformly for all axes even if
not affected by modifiers.

At the end of the arc planning, correctly set current_position to the
final logical position (that is, without modifiers).

This fixes Z interpolation in arc moves in both ABL/UBL (bug
https://github.com/prusa3d/Prusa-Firmware-Buddy/issues/3179), removing
the need for exceptions.