[FR] Sanity Check TMC_HOME_PHASE and CORE / MARKFORGED Kinematics
Core machines can't reliably settle on the same phase for homing core axes.
For instance, with CoreXY, X and Y positions depend on both A and B motors,
so the homing A phase depends on exact location of X and Y (== on the position of B motor).

Fixes https://github.com/MarlinFirmware/Marlin/issues/26307