MMU2s C0 fix:
Marlin keeps extruding filament while waiting on the "C0" to be executed by the MMU2s. Incase the MMU2s is having issues, like the 5 flashing lights or some other problems that requires manual intervention and cannot respond to requests properly and the filament is already in the extruder, a huge bulge of filament on to the wipe tower is extruded which is not recoverable anymore. This fixes that by replacing the `unscaled_mmu2_e_move` call with an `idle` call.