Create tronxy_script.py
Creates compiled fmw_tronxy.bin and fmw_tronxy.hex firmware files and places them in a folder labeled "update" in the root of the Marlin project folder for easy access for the user.
The Tronxy V10 board requires the firmware to be named in this manner and inside the folder "update" on the SD card for flashing. This saves the user a few steps.