🐛 Respect odd number of strokes when cleaning nozzle with linear pattern
Users may wish to set NOZZLE_CLEAN_STROKES to an odd number.  This can
be used to prevent a glob of plastic from re-adhering to the nozzle on
the final cleaning stroke.

Previously, "G12 P0 S2" and "G12 P0 S3" performed the same operation.
With this change, specifying an odd number of strokes causes the
cleaning sequence to end at NOZZLE_CLEAN_END_POINT, instead of
unconditionally returning to NOZZLE_CLEAN_START_POINT.